+++ process_mining/src/analysis/mod.rs
@@ -1,0 +1,6 @@
+//! Process Analysis
+//!
+//! This module contains techniques related to the analysis of event data or other process artifacts.
+
+pub mod case_centric;
+pub mod object_centric;