+++ process_mining/src/analysis/object_centric/mod.rs
@@ -1,0 +1,3 @@
+//! Object-centric Process Analysis
+
+pub mod object_attribute_changes;